Migrate Mutation.leaveOrganization
